Modernize properties/multi_properties C++
- Use range-based for/algorithms where possible
- Simplify if branches
- Order member functions following IsoCPP guidelines
- Remove definitions no longer required in C++11

Use size_t for counting in {multi}properties as signed integers cause
confusion when used as interfaces to naturally unsigned quantities
such as counting. For collections and sizing the builtin size_t type
is a more natural and safe choice, and consistent with the STL.

Replace use of int in {multi}properties interfaces with size_t.
Use max of size_t as scalar marker to avoid confusion when data
is constructed with an empty vector. When initialized with an
empty vector, size will be zero and thus scalar, if using 0 as
scalar marker. Change to use max of size_t as safe marker as certain
not to require a vector with this number of entries.

+++ source/bxdatatools/src/object_configuration_description.cc
@@ -430,7 +430,7 @@ namespace datatools {
       if (is_string()) {
         return _default_value_string_ != "__??__";
       }
-    } 
+    }
     return false;
   }
 
@@ -1565,7 +1565,7 @@ namespace datatools {
       }
       // Check fixed sized array :
       if (cpd_.is_fixed_sized_array()) {
-        if (cpd_.get_array_fixed_size() != pd.get_size()) {
+        if (cpd_.get_array_fixed_size() != (int)pd.get_size()) {
           std::ostringstream message_oss;
           message_oss << "Expected array size "
                       << cpd_.get_array_fixed_size()
